build: use esm as bin command
