+++ package.json
@@ -3,16 +3,16 @@
 	"version": "1.0.0",
 	"types": "./dist/types/index.d.ts",
 	"main": "./dist/cjs/index.js",
-	"bin": {
-		"pkgtest": "dist/bin/pkgtest.js"
-	},
+	"bin": "dist/esm/bin/pkgtest.mjs",
 	"exports": {
 		"types": "./dist/types/index.d.ts",
 		"require": "./dist/cjs/index.js",
 		"import": "./dist/esm/index.js",
 		"default": "./dist/esm/index.js"
 	},
-	"files": ["dist"],
+	"files": [
+		"dist"
+	],
 	"dependencies": {
 		"chalk": "^5.4.1",
 		"commander": "^13.1.0",
@@ -66,7 +66,7 @@
 		"format": "biome format",
 		"build:esm": "tswc -- src -d dist/esm --config-file .esm.swcrc --strip-leading-paths --out-file-extension mjs",
 		"build:cjs": "tswc -- src -d dist/cjs --config-file .cjs.swcrc --strip-leading-paths",
-		"build:bin": "rspack build",
+		"build:bin": "node ./bin-build.js",
 		"build:types": "tsc",
 		"build": "yarn build:types && yarn build:esm && yarn build:cjs && yarn build:bin",
 		"prepack": "pinst --disable",
